Fix file corruption using 32-bit Windows
The issue was caused by _fseeki64 / _ftelli64, which are completely
broken when compiling for 32-bit Windows using Mingw/Mingw-w64.
Instead, use -D_FILE_OFFSET_BITS=64 and fseeko / ftello for LFS, which
were already used on non-Windows platforms. This is supported by
Mingw-w64.

The file_pos typedef was also removed, as it was only used in an
earlier prototype.

Also, update the URL to the project's license file, which previously
referred to my old GitHub username.